examples: migrate python examples to the new client API
- python-r4-us-core: uses the default fhirpy client (no explicit option); the fhirpy
  AsyncFHIRClient demo is demo.py (renamed from client.py — it is a demo, not a Client class)
  and serializes via the generated serialize() helper
- python-r4 (renamed from python-fhirpy): the simple requests Client, generated with client: "none"
- Makefile/CI: rename the python-fhirpy targets/jobs/paths to python-r4
- READMEs: fix client descriptions, unwrap hard-wrapped paragraphs; add a Markdown no-wrap rule to CLAUDE.md
- No example uses the deprecated fhirpyClient flag anymore
