+++ examples/python-r4/generate.ts
@@ -1,17 +1,17 @@
 import { APIBuilder, prettyReport } from "../../src";
 
 if (require.main === module) {
-    console.log("📦 Generating FHIR R4 Core Types with fhirpy support...");
+    console.log("📦 Generating FHIR R4 Core Types with simple requests client...");
 
     const builder = new APIBuilder()
         .throwException()
         .fromPackage("hl7.fhir.r4.core", "4.0.1")
         .python({
             allowExtraFields: false,
-            fieldFormat: "camelCase",
-            fhirpyClient: true,
+            fieldFormat: "snake_case",
+            client: "none",
         })
-        .outputTo("./examples/python-fhirpy/fhir_types")
+        .outputTo("./examples/python-r4/fhir_types")
         .cleanOutput(true);
 
     const report = await builder.generate();
